some xml refactors
